feature/auth

+++ src/main/java/org/example/auth/CustomUserPrincipal.java
@@ -10,13 +10,13 @@ import org.springframework.security.core.userdetails.UserDetails;
 // 주로 Security 컨텍스트에 현재 사용자의 정보를 제공하는 용도로 사용됩니다.
 public class CustomUserPrincipal implements UserDetails {
 
-    private final Long userId;
+    private final Integer userId;
     private final String loginId;
     private final String username;
     private final String passwordHash;
     private final UserRole role;
 
-    public CustomUserPrincipal(Long userId, String loginId, String username, String passwordHash, UserRole role) {
+    public CustomUserPrincipal(Integer userId, String loginId, String username, String passwordHash, UserRole role) {
         this.userId = userId;
         this.loginId = loginId;
         this.username = username;
@@ -24,7 +24,7 @@ public class CustomUserPrincipal implements UserDetails {
         this.role = role;
     }
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
